Added pandoc-zotxt.lua
This is a pandoc filter that does more or less the same that
pandoc-zotxt does. I had hoped for it to be faster than the
original, but it isn't. (I wrongly assumed that I could retrieve
multiple citation items concurrently in Lua; but Pandoc doesn't
support that.) However, it ships with its only dependency, a
JSON library, and doesn't require anything to run but Pandoc
itself. So maybe it's useful for some people.